refactor(cli): lift the `pair new` frame primitives into a shared module
#364 built a framed, TTY-gated shape for `actana pair new` and put the border
width, the padding rule, the colour table and the display-width measurement
inside `actana-pair.ts`, because there was one framed block in the program.

#360 adds the second one — `actana core pair`, the client end of the very same
exchange. Two copies of a border width and a colour table is how the two ends
of one handshake start drawing different boxes, so the primitives move to
`cli-frame.ts` and both callers import them.

Nothing about the output moves. `actana-pair.ts` keeps its handout, its wording
and its one `if`; `actana-pair.test.ts` is untouched and still passes whole, so
the piped byte-identity and the framed content it pins are the proof that this
is a lift and not a rewrite. `displayWidth` and `FRAME_WIDTH` are re-exported
from `actana-pair.ts` because that suite measures the handout with them.

Two additions, neither reachable from `pair new`:

- `ANSI.green` and `ANSI.red`, the markers a result block needs and a handout
  does not;
- `wrapText`, which breaks prose on spaces by display column and — like
  `wrapFingerprint` beside it — cannot shorten its input.

Refs #360
