+++ packages/cli/src/actana-pair.ts
@@ -88,9 +88,28 @@ import {
   type PairingAuditSink,
 } from "@actana/shared/pairing-audit";
 import { readActanaConfig } from "./actana-config.ts";
+import {
+  ANSI,
+  clip,
+  displayWidth,
+  frameEdge,
+  frameRow,
+  FRAME_GUTTER,
+  FRAME_WIDTH,
+  style,
+  useColor,
+  type Span,
+} from "./cli-frame.ts";
 import { formatJson, formatTable } from "./cli-output.ts";
 import { EXIT_FAILURE, EXIT_OK, EXIT_USAGE } from "./exit-codes.ts";
 import type { ActanaCliDeps } from "./cli-deps.ts";
+
+// The frame's own vocabulary, re-exported because it is part of what this
+// module prints and `actana-pair.test.ts` measures the handout with it. The
+// definitions moved to `cli-frame.ts` when `actana core pair` grew a frame of
+// its own (#360) — there is one border width in this program, not two.
+export { displayWidth, FRAME_WIDTH } from "./cli-frame.ts";
+
 
 export const PAIR_HELP = `actana pair — enroll a client on THIS machine's Core
 
@@ -462,37 +481,6 @@ function chooseEndpointHost(
 // assert on the frame without a terminal, and what keeps the decision to draw
 // one in exactly one `if`.
 
-/** The escape sequences the frame uses, and the only ANSI in this file. */
-const ANSI = {
-  reset: "\u001b[0m",
-  bold: "\u001b[1m",
-  dim: "\u001b[2m",
-  /** The pairing code itself, which is the one thing on the screen that matters. */
-  boldCyan: "\u001b[1;36m",
-} as const;
-
-/**
- * Whether to colour, given a terminal.
- *
- * `stdoutIsTty` is the gate; `NO_COLOR` is the standard opt-out on top of it
- * (https://no-color.org). Honouring it costs one clause and means an operator
- * whose terminal renders escapes badly still gets the instructions, rather than
- * having to choose between colour and a pipe.
- */
-function useColor(deps: ActanaCliDeps): boolean {
-  const noColor = deps.env.NO_COLOR;
-  return deps.stdoutIsTty && (noColor === undefined || noColor === "");
-}
-
-/** The frame's total width, borders included. Fixed, so the output is one shape. */
-export const FRAME_WIDTH = 74;
-
-/** How far in from the left border the content starts. */
-const FRAME_GUTTER = 3;
-
-/** One run of styled text inside a framed line. */
-type Span = { text: string; style?: string };
-
 /**
  * Every fact the handout renders. Times arrive as epoch ms and the clock does
  * not: `now` is passed in, because this file reads the clock in exactly one
@@ -754,108 +742,6 @@ export function wrapFingerprint(fingerprint: string, groupsPerLine = 16): string
     lines.push(groups.slice(i, i + perLine).join(":") + (last ? "" : ":"));
   }
   return lines;
-}
-
-/** The top or bottom rule. */
-function frameEdge(which: "top" | "bottom", color: boolean): string {
-  const [left, right] = which === "top" ? ["╭", "╮"] : ["╰", "╯"];
-  return style(`${left}${"─".repeat(FRAME_WIDTH - 2)}${right}`, ANSI.dim, color);
-}
-
-/**
- * One line between the two borders.
- *
- * The padding is measured on the **plain** text and the styling applied
- * afterwards, because an escape sequence has a length and no width — pad a
- * coloured string by `String.length` and the right border walks off by however
- * many bytes the colour cost.
- */
-function frameRow(spans: Span[], color: boolean): string {
-  const plain = spans.map((span) => span.text).join("");
-  const body = spans.map((span) => style(span.text, span.style, color)).join("");
-  const fill = Math.max(1, FRAME_WIDTH - 2 - FRAME_GUTTER - displayWidth(plain));
-  const bar = style("│", ANSI.dim, color);
-  return `${bar}${" ".repeat(FRAME_GUTTER)}${body}${" ".repeat(fill)}${bar}`;
-}
-
-/**
- * How many terminal columns `text` occupies.
- *
- * `String.length` is UTF-16 code units and a frame padded by it is crooked for
- * anybody whose label is not Latin: `笔记本` is three code units and six
- * columns, so the right border lands three early (#357 review N1). Counted per
- * code point — a surrogate pair is one character, not two — with the wide
- * ranges at two and the zero-width ones at nothing.
- *
- * Deliberately not a full `wcwidth`. This measures the four things that reach
- * a frame row: ASCII, an operator's label, a UUID and colon-hex. Emoji with
- * joiners and skin tones can still measure a column off, and the cost of that
- * is a ragged border on one line — not a truncated fingerprint, which is the
- * thing this file may never do.
- */
-export function displayWidth(text: string): number {
-  let width = 0;
-  for (const character of text) {
-    const point = character.codePointAt(0) ?? 0;
-    if (isZeroWidth(point)) continue;
-    width += isWide(point) ? 2 : 1;
-  }
-  return width;
-}
-
-/** Combining marks and variation selectors: they attach, they take no column. */
-function isZeroWidth(point: number): boolean {
-  return (
-    (point >= 0x0300 && point <= 0x036f) ||
-    (point >= 0x200b && point <= 0x200f) ||
-    (point >= 0xfe00 && point <= 0xfe0f) ||
-    point === 0x2060 ||
-    point === 0xfeff
-  );
-}
-
-/** The double-width blocks: CJK, Hangul, kana, fullwidth forms, and emoji. */
-function isWide(point: number): boolean {
-  return (
-    (point >= 0x1100 && point <= 0x115f) ||
-    (point >= 0x2e80 && point <= 0xa4cf) ||
-    (point >= 0xac00 && point <= 0xd7a3) ||
-    (point >= 0xf900 && point <= 0xfaff) ||
-    (point >= 0xfe30 && point <= 0xfe6f) ||
-    (point >= 0xff00 && point <= 0xff60) ||
-    (point >= 0xffe0 && point <= 0xffe6) ||
-    (point >= 0x1f300 && point <= 0x1f64f) ||
-    (point >= 0x1f900 && point <= 0x1f9ff) ||
-    (point >= 0x20000 && point <= 0x3fffd)
-  );
-}
-
-/**
- * `text` shortened to `columns` display columns, marked when it was shortened.
- *
- * **Only ever called on the label.** Truncation is a lie about a value, and
- * there is exactly one value in this block cheap enough to tell it about: a
- * label is a name the operator chose and can read back off `pair ls`. The
- * fingerprint wraps instead — see {@link wrapFingerprint} — and nothing routes
- * it through here.
- */
-function clip(text: string, columns: number): string {
-  if (displayWidth(text) <= columns) return text;
-  let kept = "";
-  let width = 0;
-  for (const character of text) {
-    const next = width + displayWidth(character);
-    // One column held back for the marker, which is one column wide.
-    if (next > columns - 1) break;
-    kept += character;
-    width = next;
-  }
-  return `${kept}…`;
-}
-
-/** Wrap `text` in an escape sequence, or hand it back untouched. */
-function style(text: string, code: string | undefined, color: boolean): string {
-  return color && code ? `${code}${text}${ANSI.reset}` : text;
 }
 
 /**
